# go-server/go/model_digimon_info_request.go
/*
 * Digimon Service API
 *
 * 提供孵化數碼蛋與培育等數碼寶貝養成服務
 *
 * API version: 1.0.0
 * Contact: 123@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type DigimonInfoRequest struct {
	// 數碼蛋的名字
	Name string `json:"name"`
}